refactor (data): Rename experiments.py to loaders.py
Rename the file previously called experiments.py to loaders.py in order to limit confusion.
